Pull logic out of CLI
